Module Wire.Response

include S.WIRE_PAYLOAD with type t = response msg and type path := Path.t

type t = response msg

A message payload. This is typically a byte array of some kind, plus a way of attaching capabilities.

val pp : t Fmt.t

val cap_index : t -> Path.t -> int option

cap_index msg path is the capability index at path in the message (i.e the index of the capability in the attachments table).

val attachments : t -> S.attachments

val with_attachments : S.attachments -> t -> t

val bootstrap : unit -> t

The (empty) content for the reply to the bootstrap message.