Module Capnp_rpc.Message_types

This module defines the information in the messages that goes over the wire. These messages are turned into actual byte streams elsewhere.

module EmbargoId : sig ... end

module type TABLE_TYPES = sig ... end

For the unit tests it is convenient to pass in the types of table indexes. This allows the tests to make both ends of a connection, with the types matched up.

module Flip (T : TABLE_TYPES) : sig ... end

Flip T is the types for the other end of T's connection.

module Make (Core_types : S.CORE_TYPES) (Network : S.NETWORK_TYPES) (T : TABLE_TYPES) : sig ... end

This module defines the information in the messages that goes over the wire in one direction. The types are from the point of view of the sender, as in the Cap'n Proto RPC specification.

module type ENDPOINT = sig ... end

module Table_types () : sig ... end

module Endpoint (Core_types : S.CORE_TYPES) (Network_types : S.NETWORK_TYPES) (Table : TABLE_TYPES) : sig ... end