Module type Message_types.ENDPOINT

module Core_types : S.CORE_TYPES

This module defines a calling convention for invoking methods on objects. The objects could be remote, but this module doesn't define anything related to networks. These types are generated automatically from WIRE by Core_types.

module Network_types : S.NETWORK_TYPES

module Table : TABLE_TYPES

For the unit tests it is convenient to pass in the types of table indexes. This allows the tests to make both ends of a connection, with the types matched up.

module Out : sig ... end

The type of messages sent by this endpoint.

module In : sig ... end

The type of messages received by this endpoint.