Module Capnp_rpc.Service

Functions for service implementors.

type ('a, 'b) method_t = 'a -> (unit -> unit) -> 'b StructRef.t

An ('a, 'b) method_t is a method implementation that takes a reader for the parameters and a function to release the capabilities in the parameters, and returns a promise for the results.

module Response : sig ... end

val return : 'a Response.t -> 'a StructRef.t

return r wraps up a simple local result as a promise.

val return_empty : unit -> 'a StructRef.t

return_empty () is a promise for a response with no payload.

val error : Error.t -> 'a StructRef.t

error e is a broken promise for a struct, with error e.

val fail : ?ty:Exception.ty -> ('a, Stdlib.Format.formatter, unit, 'b StructRef.t) Stdlib.format4 -> 'a

fail msg is an exception error with reason msg.