Module Wire.Request

type t = request msg

A message payload. This is typically a byte array of some kind, plus a way of attaching capabilities.

type path

val pp : t Fmt.t

val cap_index : t -> path -> int option

cap_index msg path is the capability index at path in the message (i.e the index of the capability in the attachments table).

val attachments : t -> Capnp_rpc.S.attachments

val with_attachments : Capnp_rpc.S.attachments -> t -> t