Module Message_types.Make

This module defines the information in the messages that goes over the wire in one direction. The types are from the point of view of the sender, as in the Cap'n Proto RPC specification.

Parameters

module Core_types : S.CORE_TYPES

module Network : S.NETWORK_TYPES

module T : TABLE_TYPES

Signature

include module type of struct include T end

module QuestionId = T.QuestionId

module AnswerId = T.AnswerId

module ImportId = T.ImportId

module ExportId = T.ExportId

type third_party_desc = Network.third_party_cap_id * ExportId.t

type message_target = [

| `ReceiverHosted of ImportId.t

| `ReceiverAnswer of QuestionId.t * Core_types.Wire.Path.t

]

type desc = [

| message_target

| `None

| `SenderHosted of ExportId.t

| `SenderPromise of ExportId.t

| `ThirdPartyHosted of third_party_desc

]

val pp_desc : Stdlib.Format.formatter -> [< `Local of < pp : Stdlib.Format.formatter -> unit.. > | `None | `ReceiverAnswer of QuestionId.t * Core_types.Wire.Path.t | `ReceiverHosted of ImportId.t | `SenderHosted of ExportId.t | `SenderPromise of ExportId.t | `ThirdPartyHosted of 'a ] -> unit

type send_results_to = [

| `Caller

| `Yourself

| `ThirdParty of Network.recipient_id

]

type return = [

| Error.t

| `Results of Core_types.Wire.Response.t * desc RO_array.t

| `ResultsSentElsewhere

| `TakeFromOtherQuestion of QuestionId.t

| `AcceptFromThirdParty

]

type disembargo_request = [

| `Loopback of message_target * EmbargoId.t

]

val pp_return : Stdlib.Format.formatter -> [< `AcceptFromThirdParty | `Cancelled | `Exception of Exception.t | `Results of 'a * [< `Local of < pp : Stdlib.Format.formatter -> unit.. > | `None | `ReceiverAnswer of QuestionId.t * Core_types.Wire.Path.t | `ReceiverHosted of ImportId.t | `SenderHosted of ExportId.t | `SenderPromise of ExportId.t | `ThirdPartyHosted of 'b ] RO_array.t | `ResultsSentElsewhere | `TakeFromOtherQuestion of QuestionId.t ] -> unit

val pp_disembargo_request : disembargo_request Fmt.t

type t = [

| `Abort of Exception.t

| `Bootstrap of QuestionId.t * string

| `Call of QuestionId.t * message_target * Core_types.Wire.Request.t * desc RO_array.t * send_results_to

| `Finish of QuestionId.t * bool

| `Release of ImportId.t * int

| `Disembargo_request of disembargo_request

| `Disembargo_reply of message_target * EmbargoId.t

| `Return of AnswerId.t * return * bool

| `Resolve of ExportId.t * (desc, Exception.t) Stdlib.result

]

A message sent over the network.

val with_qid_tag : Logs.Tag.set -> t -> Logs.Tag.set

val pp_results_to : Stdlib.Format.formatter -> [< `Caller | `ThirdParty of 'a | `Yourself ] -> unit

val pp_recv : (Stdlib.Format.formatter -> Core_types.Wire.Request.t -> unit) -> t Fmt.t

Describe message from the point of view of the receiver.