Class type Core_types.base_ref

Common methods for struct_ref and cap.

method pp : Stdlib.Format.formatter -> unit

method update_rc : int -> unit

c#update_rc d adds d to c's (local) reference count. When it reaches zero, c must not be used again. A message may be sent releasing any remote resources.

method blocker : base_ref option

c#blocker is the unresolved cap or struct_ref promise that must resolve for c to resolve. This is used to help detect cycles.

method check_invariants : unit

This is for debugging. Checks its own invariants and those of other base_refs it holds. Raises Invariant_broken if there is a problem.

method sealed_dispatch : 'a. 'a Capnp_rpc_proto__.S.brand -> 'a option

c#sealed_dispatch brand extracts some private data of the given type.