Module Capnp_rpc

The abstract and untyped Cap'n Proto RPC protocol. Users will normally want to use the Capnp_rpc_lwt API instead, which provides a typed interface using the Cap'n Proto serialisation.

module S : sig ... end

The core types and module signatures.

module RO_array : sig ... end

A read-only array.

module Stats : sig ... end

Statistics for CapTP connections.

module Id : sig ... end

Unique identifiers.

module Debug : sig ... end

Diagnostics.

module Error : sig ... end

Errors returned by calls.

module Exception : sig ... end

Cap'n Proto exceptions.

module Message_types : sig ... end

This module defines the information in the messages that goes over the wire. These messages are turned into actual byte streams elsewhere.

module Core_types : functor (W : S.WIRE) -> S.CORE_TYPES with module Wire = W

module Local_struct_promise : sig ... end

A local promise for a call response. Queues messages locally.

module Cap_proxy : sig ... end

A local promise for a capability. Queues messages locally.

module CapTP : sig ... end

The abstract Cap'n Proto RPC network protocol.

module RC : sig ... end

A ref-count type that raises an exception on overflow.