Module Capnp_rpc

Cap'n Proto core API for defining and using services.

include module type of Capnp.BytesMessage

module Segment : sig ... end

module Message : sig ... end

module Slice : sig ... end

module StructStorage : sig ... end

module ListStorage : sig ... end

module Object : sig ... end

module Exception = Capnp_rpc_proto.Exception

module Error = Capnp_rpc_proto.Error

module StructRef : sig ... end

A promise for a response structure. You can use the generated _get_pipelined functions on a promise to get a promise for a capability inside the promise, and then pipeline messages to that promise without waiting for the response struct to arrive.

module Capability : sig ... end

A capability is a reference to an object, or to a promise for an object. You can invoke methods on a capability even while it is still only a promise.

module Sturdy_ref : sig ... end

module Service : sig ... end

Functions for service implementors.

module Std : sig ... end

Some aliases for common modules.

module Leak_handler : sig ... end

Handle references that got GC'd with a non-zero ref-count.

module Persistence : sig ... end