Module Capnp_rpc.Message_types

This module defines the information in the messages that goes over the wire. These messages are turned into actual byte streams elsewhere.

module EmbargoId : Capnp_rpc__Id.S

module type TABLE_TYPES = sig ... end

module Flip : functor (T : TABLE_TYPES) -> sig ... end

module Make : functor (Core_types : S.CORE_TYPES) -> functor (Network : S.NETWORK_TYPES) -> functor (T : TABLE_TYPES) -> sig ... end

module type ENDPOINT = sig ... end

module Table_types : functor () sig ... end

module Endpoint : functor (Core_types : S.CORE_TYPES) -> functor (Network_types : S.NETWORK_TYPES) -> functor (Table : TABLE_TYPES) -> sig ... end