Module type Message_types.TABLE_TYPES

For the unit tests it is convenient to pass in the types of table indexes. This allows the tests to make both ends of a connection, with the types matched up.

module QuestionId : Id.S

module AnswerId : Id.S

module ImportId : Id.S

module ExportId : Id.S