Module type S.WIRE_PAYLOAD

type t

A message payload. This is typically a byte array of some kind, plus a way of attaching capabilities.

type path

val pp : t Fmt.t

val cap_index : t -> path -> int option

cap_index msg path is the capability index at path in the message (i.e the index of the capability in the attachments table).

val attachments : t -> attachments

val with_attachments : attachments -> t -> t